requirements.txt
tensorflow-gpu=2.0.0-dev20190525
tensorflow_datasets